BUG FIX: .b64EncodeFile() could core dump [#73]
The reason for the core dump by rmd_b64encode_data() was that mkChar(NUL) was effectively called.  This would happen anytime bufcstr(databuf) return NULL, which could happen if it failed to grow the buffer 'databuf' as needed.  By testing for NULL, we now generate a proper out-of-memory error (instead of core dumping).

Also, rmd_b64encode_data() would allocate the initial 'databuf' buffer to be of size READ_UNIT (==1024), which is likely to be too small for most files encoded.  Since Base64 encoded output is at least as long as the input, I decided to make the initial buffer to have the same size at the input (== data_len).  Although I don't understand why, this latter change help allocating buffer memory and the function no longer runs out of memory for the large (14,787,751 bytes) example PNG file used in Issue #73.