avoid reading uninitialized values
It may look surprising that we directly delete the first
if-statement, but in fact the second if-statement itself will do
the work.

The purpose of this function is to return a NULL-terminated
string, so we need to make sure that data[size] == 0.

The first if- tests whether data[size] == 0, but this value
was uninitialized by realloc(). The second if- manually set
data[size] = 0, so it is safe to only use the second if-.